refactor(api): split countTotalUsersOnDate from user service
