+++ packages/api/src/modules/stats/stats.service.ts
@@ -11,6 +11,7 @@ import { RoleEnum } from "../../@enums/Roles";
 import UserDbo from "../user/repositories/dbo/UserDbo";
 import { isUserActif } from "../../shared/helpers/UserHelper";
 import * as DateHelper from "../../shared/helpers/DateHelper";
+import userStatsService from "../user/services/stats/user.stats.service";
 import userAssociationVisitJoiner from "./joiners/UserAssociationVisitsJoiner";
 import { UserWithAssociationVisitsEntity } from "./entities/UserWithAssociationVisitsEntity";
 import AssociationVisitEntity from "./entities/AssociationVisitEntity";
@@ -90,7 +91,7 @@ class StatsService {
 
     async getMonthlyUserNbByYear(year: number) {
         const start = firstDayOfPeriod(year);
-        const init_count = await userService.countTotalUsersOnDate(start);
+        const init_count = await userStatsService.countTotalUsersOnDate(start);
         const users = await userService.findByPeriod(start, oneYearAfterPeriod(year));
 
         const now = new Date();
