+++ packages/api/src/modules/user/services/stats/user.stats.service.ts
@@ -1,4 +1,10 @@
-export class UserStatsService {}
+import userRepository from "../../repositories/user.repository";
+
+export class UserStatsService {
+    public countTotalUsersOnDate(date, withAdmin = false) {
+        return userRepository.countTotalUsersOnDate(date, withAdmin);
+    }
+}
 
 const userStatsService = new UserStatsService();
 export default userStatsService;
