+++ packages/api/src/modules/user/user.service.test.ts
@@ -459,31 +459,6 @@ describe("User Service", () => {
         });
     });
 
-    describe("countTotalUsersOnDate()", () => {
-        const REPO_RETURN = 5;
-        const DATE = new Date();
-        const WITH_ADMIN = true;
-
-        beforeAll(() => mockedUserRepository.countTotalUsersOnDate.mockResolvedValue(REPO_RETURN));
-        afterAll(() => mockedUserRepository.countTotalUsersOnDate.mockRestore());
-
-        it("should call repo with given args", async () => {
-            await userService.countTotalUsersOnDate(DATE, WITH_ADMIN);
-            expect(mockedUserRepository.countTotalUsersOnDate).toBeCalledWith(DATE, WITH_ADMIN);
-        });
-
-        it("should call repo with default", async () => {
-            await userService.countTotalUsersOnDate(DATE);
-            expect(mockedUserRepository.countTotalUsersOnDate).toBeCalledWith(DATE, false);
-        });
-
-        it("should return repo's return value", async () => {
-            const expected = REPO_RETURN;
-            const actual = await userService.countTotalUsersOnDate(DATE);
-            expect(actual).toBe(expected);
-        });
-    });
-
     describe("isExpiredReset", () => {
         it("should return true", () => {
             const reset = {
